Python_Assignment/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
pydantic>=2.4
pydantic[email]
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-dotenv>=1.0
psycopg[binary]>=3.2
bcrypt==4.0.1